GROW-317 refactor: 레거시 코드 별도 legacy 폴더 내에 관리
