+++ app/data/celery_app/base.py
@@ -4,12 +4,9 @@ from celery import Celery
 from celery.schedules import crontab
 from dotenv import load_dotenv
 
-import app.data.custom.remove_realtime_data  # noqa: F401 > task 위치를 찾는데 필요합니다.
 import app.data.custom.validate_data  # noqa: F401 > task 위치를 찾는데 필요합니다.
-import app.data.polygon.usa_realtime_stock  # noqa: F401 > task 위치를 찾는데 필요합니다.
 import app.data.yahoo.dividend  # noqa: F401 > task 위치를 찾는데 필요합니다.
 import app.data.yahoo.index  # noqa: F401 > task 위치를 찾는데 필요합니다.
-import app.data.yahoo.realtime_stock.realtime_stock_app  # noqa: F401 > task 위치를 찾는데 필요합니다.
 import app.data.yahoo.stock  # noqa: F401 > task 위치를 찾는데 필요합니다.
 from database.enum import EnvironmentType
 
@@ -50,14 +47,6 @@ celery_task.conf.beat_schedule = {
     "stock": {
         "task": "app.data.yahoo.stock.main",
         "schedule": crontab(minute=0),
-    },
-    "usa_realtime_stock": {
-        "task": "app.data.polygon.usa_realtime_stock.main",
-        "schedule": crontab(minute=0),
-    },
-    "remove_minutely": {
-        "task": "app.data.custom.remove_realtime_data.main",
-        "schedule": crontab(hour=4, minute=0),
     },
     "validate_data": {
         "task": "app.data.custom.validate_data.main",
