+++ app/data/naver/legacy/realtime_index/realtime_index_collector_korea.py
