+++ app/data/naver/legacy/realtime_stock/korea_realtime_stock_collector.py
