+++ app/data/naver/legacy/realtime_stock/realtime_stock_app.py
@@ -3,8 +3,8 @@ import asyncio
 import ray
 
 from app.data.common.services.stock_code_file_service import StockCodeFileReader
-from app.data.naver.realtime_stock.korea_realtime_stock_collector import KoreaRealtimeStockCollector
-from app.data.naver.realtime_stock.realtime_stock_monitor import RealtimeStockMonitor
+from app.data.naver.legacy.realtime_stock.korea_realtime_stock_collector import KoreaRealtimeStockCollector
+from app.data.naver.legacy.realtime_stock.realtime_stock_monitor import RealtimeStockMonitor
 
 
 async def execute_async_task():
