+++ app/data/yahoo/legacy/realtime_stock/realtime_stock_collector.py
