--- docker/Dockerfile.realtime_stock_korea
@@ -1,20 +1,0 @@
-FROM python:3.11
-
-WORKDIR /app
-
-RUN apt-get update && \
-    apt-get install -y gcc g++ make && \
-    apt-get clean
-
-RUN pip install poetry
-RUN pip install peewee==3.17.6
-
-COPY pyproject.toml poetry.lock /app/
-
-RUN poetry config virtualenvs.create false && poetry install --no-interaction --no-ansi --no-root
-
-COPY . /app/
-
-COPY ./app/data/naver/realtime_stock_korea.py /app/task.py
-
-CMD ["python", "task.py"]
